pyproject.toml
[project]
name = "python-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "dotenv>=0.9.9",
    "mysql-connector-python>=9.4.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[tool.black]
line-length = 79

